requirements.txt
viam-sdk
pms5003
RPi.GPIO
smbus